all rake tasks working individually

+++ db/migrate/20180625232019_create_transactions.rb
@@ -1,7 +1,7 @@
 class CreateTransactions < ActiveRecord::Migration[5.2]
   def change
     create_table :transactions do |t|
-      t.integer :credit_card_number
+      t.bigint :credit_card_number
       t.integer :credit_card_expiration
       t.string :result
       t.references :invoice, foreign_key: true
